Check powerstate of switch (#5)
* Check powerstate of switch

You can check the powerstate of the switch. Pass `--powerstate on` as 
command line argument, and the script will give you a warning state when 
the Shelly switch is powered off.

* Rename variables to powerstatus

The variables were named powerstate, and powerstatus before. This was 
misleading, I think. Since they're all about the same topic, the  
variables are named powerstatus now.

The --powerstate argument was renamed to --expect-powerstatus.

* Boolean check of expect_powerstatus

* Add comment

* Exit if checktype!=meter but expect-powerstatus is given